MAINT: remove no-longer-raising docstring'd-out traceback;
Make previously-raising code into an example function, with a smoke-test to ensure it doesn't raise
